Added entity grabber
Use a carrot on a stick to "grab" and move entities. Entities snap to the grid where the player is looking, though they always stay at the same y-level as the player. Entities that can be grabbed are configured with the "can_grab" entity tag. Default is currently just armor stands.